ML/requirements.txt
numpy
pandas
rrcf
python-dotenv
google-genai
rich
requests

